gd32vf103.ld: sync with picolibc
Move the vector table to .data.init.enter and the irq and trap
entries to .text.unlikely so we no longer need to add
the special .vectors, .text.init.irq and .text.init.trap
sections.
However swap the .data.init.enter and .text.init.enter to
keep the vector table at the beginning of flash so we don't
waste 512 bytes on aligning it when picolibc puts its _start
routine in .text.init.enter.